0.3.5

+++ lib/acpi.lua
@@ -2,6 +2,8 @@
 --ACPI module
 --Advanced Configuration and Power Interface
 
+RELOADABLE = false
+
 local _shutdown = os.shutdown
 local _reboot = os.reboot
 
@@ -11,34 +13,51 @@ local __clear_temp = function()
 end
 
 local function acpi_shutdown()
-    os.debug.debug_write("[shutdown] shutting down for system halt")
-    os.debug.debug_write("[shutdown] sending SIGKILL to all processes")
-    if not os.__boot_flag then --still without proper userspace
-        os.lib.proc.__killallproc()
+    if permission.grantAccess(fs.perms.SYS) then
+        os.debug.debug_write("[shutdown] shutting down for system halt")
+        os.debug.debug_write("[shutdown] sending SIGKILL to all processes")
+        if not os.__boot_flag then --still without proper userspace
+            os.lib.proc.__killallproc()
+        end
+        os.sleep(1)
+        os.debug.debug_write("[shutdown] deleting temporary")
+        __clear_temp()
+        os.debug.debug_write("[shutdown] sending HALT.")
+        os.sleep(.5)
+        _shutdown()
+    else
+        os.ferror("acpi_shutdown: cannot shutdown without SYSTEM permission")
     end
-    os.sleep(1)
-    os.debug.debug_write("[shutdown] deleting temporary")
-    __clear_temp()
-    os.debug.debug_write("[shutdown] sending HALT.")
-    os.sleep(.5)
-    _shutdown()
+    permission.default()
 end
 
 local function acpi_reboot()
-    os.debug.debug_write("[reboot] shutting down for system reboot")
-    os.debug.debug_write("[reboot] sending SIGKILL to all processes")
-    if not os.__boot_flag then --still without proper userspace
-        os.lib.proc.__killallproc()
+    if permission.grantAccess(fs.perms.SYS) then
+        os.debug.debug_write("[reboot] shutting down for system reboot")
+        os.debug.debug_write("[reboot] sending SIGKILL to all processes")
+        if not os.__boot_flag then --still without proper userspace
+            os.lib.proc.__killallproc()
+        end
+        os.sleep(1)
+        os.debug.debug_write("[reboot] sending RBT.")
+        print('\n')
+        os.sleep(.5)
+        _reboot()
+    else
+        os.ferror("acpi_reboot: cannot reboot without SYSTEM permission")
     end
-    os.sleep(1)
-    os.debug.debug_write("[reboot] sending RBT.")
-    os.debug.debug_write('\n')
-    os.sleep(.5)
-    _reboot()
+    permission.default()
 end
 
 local function acpi_suspend()
-
+    while true do
+        term.clear()
+        term.setCursorPos(1,1)
+        local event, key = os.pullEvent('key')
+        if key ~= nil then
+            break
+        end
+    end
 end
 
 local function acpi_hibernate()
